chore: address MR comments
