+++ pyproject.toml
@@ -25,7 +25,6 @@ dependencies = [
 "cryptography (>=45.0.3)",
 "colorlog",
 "orjson (>=3.10,<4)",
-"dateparser (>=1.2.2,<2.0.0)",
 ]
 
 [project.urls]
